Deixa o service mais genérico, para ser reaproveitado em outros locais
